# Some performance improvements for graph generators