--- a/docs/src/graphtypes.md
+++ b/docs/src/graphtypes.md
@@ -11,6 +11,6 @@ In addition to providing `SimpleGraph` and `SimpleDiGraph` implementations, Ligh
 These are general guidelines to help you select the proper graph type.
 
 - In general, prefer the native `SimpleGraphs`/`SimpleDiGraphs` structures in [LightGraphs.jl](https://github.com/JuliaGraphs/LightGraphs.jl).
-- If you need edge weights and don't require large numbers of graph modifications, use [SimpleWeightedGraphs](https://github.com/JuliaGraphss/SimpleWeightedGraphs.jl).
+- If you need edge weights and don't require large numbers of graph modifications, use [SimpleWeightedGraphs](https://github.com/JuliaGraphs/SimpleWeightedGraphs.jl).
 - If you need labeling of vertices or edges, use [MetaGraphs](https://github.com/JuliaGraphs/MetaGraphs.jl).
 - If you work with very large graphs (billons to tens of billions of edges) and don't need mutability, use [StaticGraphs](https://github.com/JuliaGraphs/StaticGraphs.jl).