--- a/test/simplegraphs/generators/euclideangraphs.jl
+++ b/test/simplegraphs/generators/euclideangraphs.jl
@@ -17,6 +17,6 @@
     @test minimum(points) >= 0.
 
 
-    @test_throws ArgumentError euclidean_graph(points, L=0.01,  bc=:periodic)
+    @test_throws DomainError euclidean_graph(points, L=0.01,  bc=:periodic)
     @test_throws ArgumentError euclidean_graph(points, bc=:badbc)
 end