See #903 

So I did some changes and while there is always more to do I thought it is better to a pr now to get some feedback before I have to change too much back. (And also to get some info on code coverage).

I did some small changes to smallgraphs.jl, this does not really change anything about performance. One thing I saw was that the graphs there are all of type SimpleGraph{Int}. This could be changed to SimpleGraph{Int8}.

Then I replaced several functions in staticgraphs.jl. I'm writing directly to the adjacency list in many of these functions now. To ensure that these graphs are actually correct I wrote a function `isvalid_simplegraph` and put it in simplegraphs/runtest.jl. Not sure if this is the correct location for it though.
I also put in some checks to ensure that the eltype of a graph is large enough to store all vertices. If that it not the case it will throw an `InexactError` at the moment which might not be the right choice.

Here are some benchmarks for dense and sparse graphs:
```
@benchmark CompleteGraph(10000)

# old

BenchmarkTools.Trial: 
  memory estimate:  2.45 GiB
  allocs estimate:  140003
  --------------
  minimum time:     10.039 s (2.37% GC)
  median time:      10.039 s (2.37% GC)
  mean time:        10.039 s (2.37% GC)
  maximum time:     10.039 s (2.37% GC)
  --------------
  samples:          1
  evals/sample:     1

# new

BenchmarkTools.Trial: 
  memory estimate:  763.78 MiB
  allocs estimate:  20003
  --------------
  minimum time:     75.050 ms (0.00% GC)
  median time:      260.585 ms (69.06% GC)
  mean time:        230.154 ms (62.08% GC)
  maximum time:     357.283 ms (75.80% GC)
  --------------
  samples:          23
  evals/sample:     1
```

```
@benchmark CycleGraph(100000)
# old

BenchmarkTools.Trial: 
  memory estimate:  12.97 MiB
  allocs estimate:  200003
  --------------
  minimum time:     5.682 ms (0.00% GC)
  median time:      7.287 ms (16.07% GC)
  mean time:        7.640 ms (21.00% GC)
  maximum time:     66.630 ms (87.64% GC)
  --------------
  samples:          655
  evals/sample:     1


# new 

BenchmarkTools.Trial: 
  memory estimate:  9.92 MiB
  allocs estimate:  100003
  --------------
  minimum time:     2.316 ms (0.00% GC)
  median time:      2.556 ms (0.00% GC)
  mean time:        3.135 ms (22.99% GC)
  maximum time:     61.516 ms (90.73% GC)
  --------------
  samples:          1592
  evals/sample:     1

```
I also made some small changes to eudlideangrapjs.jl and randgraphs.jl. This leads to some performance improvements but this is mainly because I replaced `Set{SimpleEdge}` with `Set{SimpleEdge{T}}` in several places, as collections of concrete types are much faster.
Benchmark:

```
@benchmark euclidean_graph(1000, 2)

# old

BenchmarkTools.Trial: 
  memory estimate:  235.36 MiB
  allocs estimate:  3176748
  --------------
  minimum time:     630.260 ms (40.15% GC)
  median time:      680.507 ms (41.12% GC)
  mean time:        705.465 ms (45.13% GC)
  maximum time:     796.383 ms (51.64% GC)
  --------------
  samples:          8
  evals/sample:     1

# new

BenchmarkTools.Trial: 
  memory estimate:  232.15 MiB
  allocs estimate:  2590544
  --------------
  minimum time:     274.175 ms (21.92% GC)
  median time:      277.335 ms (21.88% GC)
  mean time:        284.337 ms (22.58% GC)
  maximum time:     331.183 ms (32.25% GC)
  --------------
  samples:          18
  evals/sample:     1
```